refactor(util): lift OpaquePreviewBlendCallback to Util::Subrect
ScreenshotFeature and (next commit) FoveatedRender both need the same
opaque-RGB blend state callback around Subrect::DrawEditor's Image draw
— without it, ImGui's default SRC_ALPHA blend interprets the source RT's
non-1 alpha as transparency and the preview looks like a transparency
mask instead of the rendered image. Two copies of the same helper is one
copy too many.

Move to Util::Subrect::OpaquePreviewBlendCallback. Lives in a separate
TU (Subrect_PreviewBlend.cpp) so the unit-test target — which pulls
Subrect.cpp standalone without PCH — doesn't need to link the d3d
singletons. ScreenshotFeature now calls the shared helper; FoveatedRender
adoption lands in the next commit alongside its preview-source fix.

Doc the regression risks in the header so future implementers don't
reinvent it wrong: BlendEnable=FALSE and WriteMask must exclude alpha
(VR menu plate recomposition reads back the alpha channel).
